+++ src/main/java/com/tuwaiq/capstone2_mentoringsystem/Service/EnrollmentService.java
@@ -1,9 +1,11 @@
 package com.tuwaiq.capstone2_mentoringsystem.Service;
 
 import com.tuwaiq.capstone2_mentoringsystem.Models.Course;
+import com.tuwaiq.capstone2_mentoringsystem.Models.CourseSession;
 import com.tuwaiq.capstone2_mentoringsystem.Models.Enrollment;
 import com.tuwaiq.capstone2_mentoringsystem.Models.User;
 import com.tuwaiq.capstone2_mentoringsystem.Repository.CourseRepository;
+import com.tuwaiq.capstone2_mentoringsystem.Repository.CourseSessionRepository;
 import com.tuwaiq.capstone2_mentoringsystem.Repository.EnrollmentRepository;
 import com.tuwaiq.capstone2_mentoringsystem.Repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -19,33 +21,48 @@ public class EnrollmentService {
     private final EnrollmentRepository enrollmentRepository;
     private final UserRepository userRepository;
     private final CourseRepository courseRepository;
+    private final CourseSessionRepository courseSessionRepository;
 
-    public String addEnrollment(Enrollment enrollment){
+    public String addEnrollment(Enrollment enrollment) {
         User user = userRepository.findUserById(enrollment.getUserId());
         Course course = courseRepository.findCourseById(enrollment.getCourseId());
-        if (user==null){
+        CourseSession courseSession = courseSessionRepository.findCourseSessionById(enrollment.getCourseSessionId());
+        if (user == null) {
             return "user id error";
         }
-        if (course==null){
+        if (course == null) {
             return "course id error";
         }
-        enrollment.setStatus("pending");
-        enrollment.setEnrollmentDate(LocalDateTime.now());
-        enrollmentRepository.save(enrollment);
-        return "ok";
+        if (course.getAdminStatus().equalsIgnoreCase("pending")){
+            return "course status error";
+        }
+        if (courseSession == null) {
+            return "course session id error";
+        }
+        if (courseSession.getOccupied()){
+            return "course session occupied";
+        }else {
+            enrollment.setStatus("pending");
+            enrollment.setEnrollmentDate(LocalDateTime.now());
+            enrollmentRepository.save(enrollment);
+            return "ok";
+        }
     }
 
-    public List<Enrollment> getEnrollments(){
+    public List<Enrollment> getEnrollments() {
         return enrollmentRepository.findAll();
     }
 
-    public Boolean deleteEnrollment(Integer id){
+    public String deleteEnrollment(Integer userId, Integer id) {
         Enrollment enrollment = enrollmentRepository.findEnrollmentById(id);
-        if (enrollment==null){
-            return false;
-        }else {
+        if (enrollment == null) {
+            return "enrollment id error";
+        }
+        if (!userId.equals(enrollment.getUserId())) {
+            return "user id mismatch";
+        } else {
             enrollmentRepository.delete(enrollment);
-            return true;
+            return "ok";
         }
     }
 }
